% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_salary.R
\name{annual_salary}
\alias{annual_salary}
\title{Annual salary calculator}
\usage{
annual_salary(hourly_salary, hours_per_week)
}
\arguments{
\item{hourly_salary}{a numeric value indicating the hourly salary}

\item{hours_per_week}{a numeric value indicating the total number of hours worked per week}
}
\value{

}
\description{
Annual salary calculator
}
