% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/einstein.R
\name{einstein}
\alias{einstein}
\title{E = mc^2}
\usage{
einstein(mass, speed_of_light)
}
\arguments{
\item{mass}{a numeric value indicating the mass in kilograms}

\item{speed_of_light}{a numeric value indicating the speed of light in meters per second}
}
\value{

}
\description{
E = mc^2
}
