% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_ate.R
\name{i_ate}
\alias{i_ate}
\title{Describes what the user ate for a specific meal.}
\usage{
i_ate(food, meal, time)
}
\arguments{
\item{food}{a string that indicates what food was eaten.}

\item{meal}{a string that indicates whether this food was eaten as 'breakfast', 'lunch', or 'dinner'.}

\item{time}{a number that indicates the time of day that the meal was eaten.}
}
\value{

}
\description{
Contains one function, `i_ate`, that prints a statement about the user.
}
